refactor(tools): ai-policies-post-processor and plugins-post-processor to share the same code
