+++ tools/ai-policies-post-processor/run.js
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
 
-import fs from "fs";
 import path from "path";
 import { fileURLToPath } from "url";
 import minimist from "minimist";
+import { runSchemaPostProcessor } from "../lib/schema-post-processor.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -79,58 +79,16 @@ function processDescriptions(obj) {
 
   try {
     const absoluteSchemasPath = path.resolve(__dirname, schemasPath);
-
-    if (!fs.existsSync(absoluteSchemasPath)) {
-      throw new Error(`Schemas directory not found: ${absoluteSchemasPath}`);
-    }
-
     const outputDir = path.resolve(
       __dirname,
       "../../app/_schemas/ai-gateway/policies",
     );
 
-    if (!fs.existsSync(outputDir)) {
-      fs.mkdirSync(outputDir, { recursive: true });
-      console.log(`Created output directory: ${outputDir}`);
-    }
-
-    const files = fs.readdirSync(absoluteSchemasPath);
-    const jsonFiles = files.filter((file) => file.endsWith(".json"));
-
-    if (jsonFiles.length === 0) {
-      console.log("No JSON files found in the schemas directory.");
-      return;
-    }
-
-    console.log(`Processing ${jsonFiles.length} JSON schema files...`);
-
-    let hadErrors = false;
-
-    for (const file of jsonFiles) {
-      try {
-        const inputFilePath = path.join(absoluteSchemasPath, file);
-        const outputFilePath = path.join(outputDir, file);
-
-        console.log(`Processing: ${file}`);
-
-        const schemaContent = fs.readFileSync(inputFilePath, "utf8");
-        const schema = JSON.parse(schemaContent);
-
-        const processedSchema = processDescriptions(schema);
-
-        fs.writeFileSync(
-          outputFilePath,
-          JSON.stringify(processedSchema, null, 2),
-        );
-
-        console.log(`Processed and saved: ${file} ✓`);
-      } catch (error) {
-        console.error(`Error processing file ${file}:`, error.message);
-        hadErrors = true;
-      }
-    }
-
-    console.log(`\nCompleted processing. Output saved to: ${outputDir}`);
+    const { hadErrors } = runSchemaPostProcessor({
+      absoluteSchemasPath,
+      outputDir,
+      processSchema: processDescriptions,
+    });
 
     if (hadErrors) {
       process.exit(1);
